made map page server rendered by moving client rendered mapgl component to separate module and importing into client rendered page
